+++ GUI/MainDialogs.cs
@@ -39,7 +39,7 @@ namespace CKAN
 
         //Ugly Hack. Possible fix is to alter the relationship provider so we can use a loop
         //over reason for to find a user requested mod. Or, you know, pass in a handler to it.
-        private readonly ConcurrentStack<GUIMod> last_mod_to_have_install_toggled = new ConcurrentStack<GUIMod>();
+        private readonly ConcurrentStack<GUIMod> lastModToggled = new ConcurrentStack<GUIMod>();
         public async Task<CkanModule> TooManyModsProvide(TooManyModsProvideKraken kraken)
         {
             //We want LMtHIT to be the last user selection. If we alter this handling a too many provides
@@ -57,8 +57,8 @@ namespace CKAN
 
             if (module == null)
             {
-                last_mod_to_have_install_toggled.TryPeek(out mod);
-                MarkModForInstall(mod.Identifier,uncheck:true);
+                lastModToggled.TryPeek(out mod);
+                MarkModForInstall(mod.Identifier, true);
             }
             Util.Invoke(this, () =>
             {
@@ -72,7 +72,7 @@ namespace CKAN
             if(module!=null)
                 MarkModForInstall(module.identifier);
 
-            last_mod_to_have_install_toggled.TryPop(out mod);
+            lastModToggled.TryPop(out mod);
             return module;
         }
     }